+++ contrib/development-tools/bots-feeds.sh
@@ -1,0 +1,13 @@
+#!/bin/bash
+
+echo "Bots:"
+jq '.Collector | keys | length' intelmq/bots/BOTS
+jq '.Parser | keys | length' intelmq/bots/BOTS
+jq '.Expert | keys | length' intelmq/bots/BOTS
+jq '.Output | keys | length' intelmq/bots/BOTS
+
+echo "Feeds:"
+egrep -c '^    [^ ]' intelmq/etc/feeds.yaml
+echo "Shadowserver:"
+python3 -c "import intelmq.bots.parsers.shadowserver.config; print(len(intelmq.bots.parsers.shadowserver.config.mapping))"
+